Renamed "encoding" state to "processing" state (no confusion anymore between encoding(s) => VideoEncoding instance(s) and Video state, processing => Video delegated state and VideoEncoding state)
Added reprocessing? (true when any of the video encodings is processing and has already a file) and processing? (true when any of the video encodings is processing) to Video and reprocessing? (true when an encoding is processing and has already a file)

Added a text to inform that a video is currently reprocessing in the video view